+++ packages/components/src/components/form/fieldset-wrapper/fieldset-wrapper-styles.ts
@@ -1,10 +1,11 @@
-import type { FormState, Theme } from '../../../types';
+import type { Theme } from '../../../types';
 import type { FieldsetWrapperLabelSize } from './fieldset-wrapper-utils';
 import { getCss, mergeDeep } from '../../../utils';
 import { addImportantToRule, getThemedColors } from '../../../styles';
 import { headingXSmall, headingSmall, spacing } from '@porsche-design-system/utilities-v2';
 import { getFunctionalComponentRequiredStyles } from '../../common/required/required-styles';
 import { getFunctionalComponentStateMessageStyles } from '../../common/state-message/state-message-styles';
+import type { FormState } from '../form-state';
 
 const { medium: spacingMedium } = spacing;