+++ packages/components/CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 #### Added
 
 - Validation of properties for all components
+- `Text Field Wrapper` with `input type="search"` is clearable via Escape key and custom clear button across browsers
 
 ### [2.14.0-beta.0] - 2022-07-12